feat: working trino io manager
